Update normal-variation.js
What to do next:
- Fix the README
- Start the opposite and no see variation

- Opposite day / Hunting season:
Either the puck moves in opposite direction of the user, or the target moves while you chase it. I don't know which I want to do yet.

- No see, only hear:
You can only see the user and puck, not the target. Sound on because the music gets louder as you get closer to the target (Hot/Cold type vibe)