I can't do this
- The timer is not resetting when you escape and go back in
- You have to click "n" twice to start the normal variation
- The game over screen is not appearing when the timer gets to zero
- The NHL team names are not appearing, it just says "object object"

What to do next:
- Fix the README
- Start the opposite and no see variation

- Opposite day / Hunting season:
Either the puck moves in opposite direction of the user, or the target moves while you chase it. I don't know which I want to do yet. (Maybe not this idea, something else maybe?)

- No see, only hear:
You can only see the user and puck, not the target. Sound on because the music (fix you, Habs section - Coldplay) gets louder as you get closer to the target (Hot/Cold type vibe)