Figured out last variation
I'm gonna do a variation where the goal shrinks every frame and it only resets when the puck gets to it. No timer, game ends when the goal is gone.

I keep getting errors saying that my shrinkSetup and shrinkDraw are not defined, I don't really know what the problem is though